meter: Rename /view/ route to /contract/
